Refactoring the teams to separate router
